--- a/Makefile
+++ b/Makefile
@@ -37,7 +37,7 @@ clean:
 compile_all: clean compile_test compile_debug compile_profile compile_release
 
 compile_test:
-	mkdir bin
+	@echo "make a 'bin' folder first before compiling test"
 	g++ $(FLAGS) -g ./src/AES.cpp ./tests/tests.cpp -D CLASSIC_MAKE -lgtest -lpthread -o bin/test
 
 compile_debug:
@@ -62,5 +62,5 @@ run_release:
 	bin/release
 
 run_clean:
-	rm -rf bin 
+	rm -rf bin
 	mkdir bin -p