--- a/src/AES.cpp
+++ b/src/AES.cpp
@@ -21,7 +21,13 @@ AES::AES(int keyLen)
     throw "Incorrect key length";
   }
 
+  RoundedKeys = NULL;
   blockBytesLen = 4 * this->Nb * sizeof(unsigned char);
+}
+
+AES::~AES()
+{
+  if(RoundedKeys!=NULL) delete [] RoundedKeys;
 }
 
 
@@ -43,6 +49,21 @@ unsigned char * AES::EncryptECB(unsigned char in[], unsigned int inLen, unsigned
   return out;
 }
 
+unsigned char * AES::EncryptECB(unsigned char in[], unsigned int inLen, unsigned int &outLen)
+{
+  outLen = GetPaddingLength(inLen);
+  unsigned char *alignIn  = PaddingNulls(in, inLen, outLen);
+  unsigned char *out = new unsigned char[outLen];
+  for (unsigned int i = 0; i < outLen; i+= blockBytesLen)
+  {
+    EncryptBlock(alignIn + i, out + i, RoundedKeys);
+  }
+  
+  delete[] alignIn;
+  
+  return out;
+}
+
 unsigned char * AES::DecryptECB(unsigned char in[], unsigned int inLen, unsigned  char key[])
 {
   unsigned char *out = new unsigned char[inLen];
@@ -58,6 +79,15 @@ unsigned char * AES::DecryptECB(unsigned char in[], unsigned int inLen, unsigned
   return out;
 }
 
+unsigned char * AES::DecryptECB(unsigned char in[], unsigned int inLen)
+{
+  unsigned char *out = new unsigned char[inLen];
+  for (unsigned int i = 0; i < inLen; i+= blockBytesLen)
+  {
+    DecryptBlock(in + i, out + i, RoundedKeys);
+  }  
+  return out;
+}
 
 unsigned char *AES::EncryptCBC(unsigned char in[], unsigned int inLen, unsigned  char key[], unsigned char * iv, unsigned int &outLen)
 {
@@ -82,6 +112,26 @@ unsigned char *AES::EncryptCBC(unsigned char in[], unsigned int inLen, unsigned 
   return out;
 }
 
+unsigned char *AES::EncryptCBC(unsigned char in[], unsigned int inLen, unsigned char * iv, unsigned int &outLen)
+{
+  outLen = GetPaddingLength(inLen);
+  unsigned char *alignIn  = PaddingNulls(in, inLen, outLen);
+  unsigned char *out = new unsigned char[outLen];
+  unsigned char *block = new unsigned char[blockBytesLen];
+  memcpy(block, iv, blockBytesLen);
+  for (unsigned int i = 0; i < outLen; i+= blockBytesLen)
+  {
+    XorBlocks(block, alignIn + i, block, blockBytesLen);
+    EncryptBlock(block, out + i, RoundedKeys);
+    memcpy(block, out + i, blockBytesLen);
+  }
+  
+  delete[] block;
+  delete[] alignIn;
+
+  return out;
+}
+
 unsigned char *AES::DecryptCBC(unsigned char in[], unsigned int inLen, unsigned  char key[], unsigned char * iv)
 {
   unsigned char *out = new unsigned char[inLen];
@@ -98,6 +148,23 @@ unsigned char *AES::DecryptCBC(unsigned char in[], unsigned int inLen, unsigned 
   
   delete[] block;
   delete[] roundKeys;
+
+  return out;
+}
+
+unsigned char *AES::DecryptCBC(unsigned char in[], unsigned int inLen, unsigned char * iv)
+{
+  unsigned char *out = new unsigned char[inLen];
+  unsigned char *block = new unsigned char[blockBytesLen];
+  memcpy(block, iv, blockBytesLen);
+  for (unsigned int i = 0; i < inLen; i+= blockBytesLen)
+  {
+    DecryptBlock(in + i, out + i, RoundedKeys);
+    XorBlocks(block, out + i, out + i, blockBytesLen);
+    memcpy(block, in + i, blockBytesLen);
+  }
+  
+  delete[] block;
 
   return out;
 }
@@ -127,6 +194,28 @@ unsigned char *AES::EncryptCFB(unsigned char in[], unsigned int inLen, unsigned 
   return out;
 }
 
+unsigned char *AES::EncryptCFB(unsigned char in[], unsigned int inLen, unsigned char * iv, unsigned int &outLen)
+{
+  outLen = GetPaddingLength(inLen);
+  unsigned char *alignIn  = PaddingNulls(in, inLen, outLen);
+  unsigned char *out = new unsigned char[outLen];
+  unsigned char *block = new unsigned char[blockBytesLen];
+  unsigned char *encryptedBlock = new unsigned char[blockBytesLen];
+  memcpy(block, iv, blockBytesLen);
+  for (unsigned int i = 0; i < outLen; i+= blockBytesLen)
+  {
+    EncryptBlock(block, encryptedBlock, RoundedKeys);
+    XorBlocks(alignIn + i, encryptedBlock, out + i, blockBytesLen);
+    memcpy(block, out + i, blockBytesLen);
+  }
+  
+  delete[] block;
+  delete[] encryptedBlock;
+  delete[] alignIn;
+
+  return out;
+}
+
 unsigned char *AES::DecryptCFB(unsigned char in[], unsigned int inLen, unsigned  char key[], unsigned char * iv)
 {
   unsigned char *out = new unsigned char[inLen];
@@ -145,6 +234,25 @@ unsigned char *AES::DecryptCFB(unsigned char in[], unsigned int inLen, unsigned 
   delete[] block;
   delete[] encryptedBlock;
   delete[] roundKeys;
+
+  return out;
+}
+
+unsigned char *AES::DecryptCFB(unsigned char in[], unsigned int inLen, unsigned char * iv)
+{
+  unsigned char *out = new unsigned char[inLen];
+  unsigned char *block = new unsigned char[blockBytesLen];
+  unsigned char *encryptedBlock = new unsigned char[blockBytesLen];
+  memcpy(block, iv, blockBytesLen);
+  for (unsigned int i = 0; i < inLen; i+= blockBytesLen)
+  {
+    EncryptBlock(block, encryptedBlock, RoundedKeys);
+    XorBlocks(in + i, encryptedBlock, out + i, blockBytesLen);
+    memcpy(block, in + i, blockBytesLen);
+  }
+  
+  delete[] block;
+  delete[] encryptedBlock;
 
   return out;
 }
@@ -270,25 +378,25 @@ void AES::SubBytes(unsigned char **state)
       state[i][j] = sbox[t / 16][t % 16];
     }
   }
-
-}
-
-void AES::ShiftRow(unsigned char **state, int i, int n)    // shift row i on n positions
-{
-  unsigned char *tmp = new unsigned char[Nb];
-  for (int j = 0; j < Nb; j++) {
-    tmp[j] = state[i][(j + n) % Nb];
-  }
-  memcpy(state[i], tmp, Nb * sizeof(unsigned char));
-	
-  delete[] tmp;
 }
 
 void AES::ShiftRows(unsigned char **state)
 {
-  ShiftRow(state, 1, 1);
-  ShiftRow(state, 2, 2);
-  ShiftRow(state, 3, 3);
+  //row 2
+  unsigned char buffer = state[1][0];
+  memmove(state[1],state[1]+1,sizeof(unsigned char)*3);
+  state[1][3] = buffer;
+
+  // row 3
+  unsigned char thrid[2];
+  memcpy(thrid,state[2],sizeof(unsigned char)*2);
+  memcpy(state[2],state[2]+2,sizeof(unsigned char)*2);
+  memcpy(state[2]+2,thrid,sizeof(unsigned char)*2);
+
+  // row 4
+  buffer = state[3][3];
+  memmove(state[3]+1,state[3],sizeof(unsigned char)*3);
+  state[3][0] = buffer;
 }
 
 unsigned char AES::xtime(unsigned char b)    // multiply on x
@@ -296,51 +404,33 @@ unsigned char AES::xtime(unsigned char b)    // multiply on x
   return (b << 1) ^ (((b >> 7) & 1) * 0x1b);
 }
 
-
-
-/* Implementation taken from https://en.wikipedia.org/wiki/Rijndael_mix_columns#Implementation_example */
-void AES::MixSingleColumn(unsigned char *r) 
-{
-  unsigned char a[4];
-  unsigned char b[4];
-  unsigned char c;
-  unsigned char h;
-  /* The array 'a' is simply a copy of the input array 'r'
-  * The array 'b' is each element of the array 'a' multiplied by 2
-  * in Rijndael's Galois field
-  * a[n] ^ b[n] is element n multiplied by 3 in Rijndael's Galois field */ 
-  for(c=0;c<4;c++) 
-  {
-    a[c] = r[c];
-    /* h is 0xff if the high bit of r[c] is set, 0 otherwise */
-    h = (unsigned char)((signed char)r[c] >> 7); /* arithmetic right shift, thus shifting in either zeros or ones */
-    b[c] = r[c] << 1; /* implicitly removes high bit because b[c] is an 8-bit char, so we xor by 0x1b and not 0x11b in the next line */
-    b[c] ^= 0x1B & h; /* Rijndael's Galois field */
-  }
-  r[0] = b[0] ^ a[3] ^ a[2] ^ b[1] ^ a[1]; /* 2 * a0 + a3 + a2 + 3 * a1 */
-  r[1] = b[1] ^ a[0] ^ a[3] ^ b[2] ^ a[2]; /* 2 * a1 + a0 + a3 + 3 * a2 */
-  r[2] = b[2] ^ a[1] ^ a[0] ^ b[3] ^ a[3]; /* 2 * a2 + a1 + a0 + 3 * a3 */
-  r[3] = b[3] ^ a[2] ^ a[1] ^ b[0] ^ a[0]; /* 2 * a3 + a2 + a1 + 3 * a0 */
-}
-
-/* Performs the mix columns step. Theory from: https://en.wikipedia.org/wiki/Advanced_Encryption_Standard#The_MixColumns_step */
 void AES::MixColumns(unsigned char** state) 
 {
-  unsigned char *temp = new unsigned char[4];
+  unsigned char temp_state[4][4]; // <- please don't put this on heap if there is no error at the first place, because its faster this way
 
-  for(int i = 0; i < 4; ++i)
+  for(size_t i=0; i<4; ++i)
   {
-    for(int j = 0; j < 4; ++j)
-    {
-      temp[j] = state[j][i]; //place the current state column in temp
-    }
-    MixSingleColumn(temp); //mix it using the wiki implementation
-    for(int j = 0; j < 4; ++j)
-    {
-      state[j][i] = temp[j]; //when the column is mixed, place it back into the state
-    }
+    memset(temp_state[i],0,4);
   }
-  delete[] temp;
+
+  for(size_t i=0; i<4; ++i)
+  {
+    for(size_t k=0; k<4; ++k)
+    {
+      for(size_t j=0; j<4; ++j)
+      {
+        if(CMDS[i][k]==1)
+          temp_state[i][j] ^= state[k][j];
+        else
+          temp_state[i][j] ^= GF_MUL_TABLE[CMDS[i][k]][state[k][j]];
+        }
+      }
+  }
+
+  for(size_t i=0; i<4; ++i)
+  {
+    memcpy(state[i],temp_state[i],4);
+  }
 }
 
 void AES::AddRoundKey(unsigned char **state, unsigned char *key)
@@ -436,9 +526,53 @@ void AES::KeyExpansion(unsigned char key[], unsigned char w[])
 
   delete []rcon;
   delete []temp;
-
 }
 
+void AES::KeyExpansion(unsigned char key[])
+{
+  unsigned char *w = new unsigned char[4 * Nb * (Nr + 1)];
+  unsigned char *temp = new unsigned char[4];
+  unsigned char *rcon = new unsigned char[4];
+
+  int i = 0;
+  while (i < 4 * Nk)
+  {
+    w[i] = key[i];
+    i++;
+  }
+
+  i = 4 * Nk;
+  while (i < 4 * Nb * (Nr + 1))
+  {
+    temp[0] = w[i - 4 + 0];
+    temp[1] = w[i - 4 + 1];
+    temp[2] = w[i - 4 + 2];
+    temp[3] = w[i - 4 + 3];
+
+    if (i / 4 % Nk == 0)
+    {
+        RotWord(temp);
+        SubWord(temp);
+        Rcon(rcon, i / (Nk * 4));
+        XorWords(temp, rcon, temp);
+    }
+    else if (Nk > 6 && i / 4 % Nk == 4)
+    {
+      SubWord(temp);
+    }
+
+    w[i + 0] = w[i - 4 * Nk] ^ temp[0];
+    w[i + 1] = w[i + 1 - 4 * Nk] ^ temp[1];
+    w[i + 2] = w[i + 2 - 4 * Nk] ^ temp[2];
+    w[i + 3] = w[i + 3 - 4 * Nk] ^ temp[3];
+    i += 4;
+  }
+
+  delete []rcon;
+  delete []temp;
+
+  RoundedKeys = w;
+}
 
 void AES::InvSubBytes(unsigned char **state)
 {
@@ -454,60 +588,49 @@ void AES::InvSubBytes(unsigned char **state)
   }
 }
 
-
-unsigned char AES::mul_bytes(unsigned char a, unsigned char b) // multiplication a and b in galois field
-{
-    unsigned char p = 0;
-    unsigned char high_bit_mask = 0x80;
-    unsigned char high_bit = 0;
-    unsigned char modulo = 0x1B; /* x^8 + x^4 + x^3 + x + 1 */
-
-
-    for (int i = 0; i < 8; i++) {
-      if (b & 1) {
-           p ^= a;
-      }
-
-      high_bit = a & high_bit_mask;
-      a <<= 1;
-      if (high_bit) {
-          a ^= modulo;
-      }
-      b >>= 1;
-    }
-
-    return p;
-}
-
-
 void AES::InvMixColumns(unsigned char **state)
 {
-  unsigned char s[4], s1[4];
-  int i, j;
+  unsigned char temp_state[4][4]; // <- please don't put this on heap if there is no error at the first place, because its faster this way
 
-  for (j = 0; j < Nb; j++)
+  for(size_t i=0; i<4; ++i)
   {
-    for (i = 0; i < 4; i++)
-    {
-      s[i] = state[i][j];
-    }
-    s1[0] = mul_bytes(0x0e, s[0]) ^ mul_bytes(0x0b, s[1]) ^ mul_bytes(0x0d, s[2]) ^ mul_bytes(0x09, s[3]);
-    s1[1] = mul_bytes(0x09, s[0]) ^ mul_bytes(0x0e, s[1]) ^ mul_bytes(0x0b, s[2]) ^ mul_bytes(0x0d, s[3]);
-    s1[2] = mul_bytes(0x0d, s[0]) ^ mul_bytes(0x09, s[1]) ^ mul_bytes(0x0e, s[2]) ^ mul_bytes(0x0b, s[3]);
-    s1[3] = mul_bytes(0x0b, s[0]) ^ mul_bytes(0x0d, s[1]) ^ mul_bytes(0x09, s[2]) ^ mul_bytes(0x0e, s[3]);
+    memset(temp_state[i],0,4);
+  }
 
-    for (i = 0; i < 4; i++)
+  for(size_t i=0; i<4; ++i)
+  {
+    for(size_t k=0; k<4; ++k)
     {
-      state[i][j] = s1[i];
+      for(size_t j=0; j<4; ++j)
+      {
+          temp_state[i][j] ^= GF_MUL_TABLE[INV_CMDS[i][k]][state[k][j]];
+      }
     }
+  }
+
+  for(size_t i=0; i<4; ++i)
+  {
+    memcpy(state[i],temp_state[i],4);
   }
 }
 
 void AES::InvShiftRows(unsigned char **state)
 {
-  ShiftRow(state, 1, Nb - 1);
-  ShiftRow(state, 2, Nb - 2);
-  ShiftRow(state, 3, Nb - 3);
+  //row 2
+  unsigned char buffer = state[1][3];
+  memmove(state[1]+1,state[1],sizeof(unsigned char)*3);
+  state[1][0] = buffer;
+
+  // row 3
+  unsigned char thrid[2];
+  memcpy(thrid,state[2],sizeof(unsigned char)*2);
+  memcpy(state[2],state[2]+2,sizeof(unsigned char)*2);
+  memcpy(state[2]+2,thrid,sizeof(unsigned char)*2);
+
+  // row 4
+  buffer = state[3][0];
+  memmove(state[3],state[3]+1,sizeof(unsigned char)*3);
+  state[3][3] = buffer;
 }
 
 void AES::XorBlocks(unsigned char *a, unsigned char * b, unsigned char *c, unsigned int len)