@SergeyBel 2021-11-17
* optimized InvMixColumns() and micro-optimized MixColumns() - may be useful
* optimized ShiftRows()and optimized InvShiftRows() - this "optimization" has no noticeable improvement effect, but it greatly reduces readability of code. it does more harm than good
* added one type of overload for each modes of AES...- maybe idea makes sense but implementation is not good. Calls like `aes.KeyExpansion(key);` very inconvenient and the code itself is in bad style: a lot of code duplication and etc. I think it's not really needed yet

Moreover:
* Please don't modify Makefile and dev "infrastructure" except in obvious situations like typos
* There are a lot of useless comments like `<- please don't put this on heap...`  - they are not useful, only distracting

To summarize:
optimization of InvMixColumns() and MixColumns() looks interesting (I have not analyzed it in detail) and can be proposed in a separate PR. Rest changes are not needed